+++ plugins/main/common/constants.ts
@@ -634,6 +634,8 @@ export const UI_COLOR_STATUS = {
   disabled: '#646A77',
   info: '#6092C0',
   default: '#000000',
+  failed: '#CC5642',
+  notApplicable: '#B9A888',
 } as const;
 
 export const API_NAME_AGENT_STATUS = {
@@ -808,24 +810,6 @@ export enum HTTP_STATUS_CODES {
   INSUFFICIENT_STORAGE = 507,
   NETWORK_AUTHENTICATION_REQUIRED = 511,
 }
-
-export const MODULE_SCA_CHECK_RESULT_LABEL: Record<
-  Uppercase<string>,
-  { key: string; value: string }
-> = {
-  PASSED: {
-    key: 'passed',
-    value: 'Passed',
-  },
-  FAILED: {
-    key: 'failed',
-    value: 'Failed',
-  },
-  NOT_RUN: {
-    key: 'not run',
-    value: 'Not run',
-  },
-} as const;
 
 // Search bar