+++ plugins/main/server/lib/sample-data/dataset/states-sca/main.js
@@ -1,7 +1,15 @@
+const {
+  CheckResult,
+} = require('../../../../../public/components/overview/sca/utils/constants');
 const random = require('../../lib/random');
 const { generateRandomAgent, generateRandomWazuh } = require('../shared-utils');
 
-const scaResults = ['passed', 'failed', 'Not run'];
+const scaResults = [
+  CheckResult.Passed,
+  CheckResult.Failed,
+  CheckResult.NotRun,
+  CheckResult.NotApplicable,
+];
 
 function word() {
   const words = ['network', 'smb', 'auth', 'firewall', 'admin', 'crypto'];